Refactor
Move tests out of model_service
Remove unused imports
Rename packages in application layer
