+++ .gitignore
@@ -1,3 +1,4 @@
+.pytest_cache/
 __pycache__/
 .idea/
 .DS_store
