+++ model_service/application/commands/start_ffn_training.py
