--- model_service/tests/conftest.py
@@ -1,17 +1,0 @@
-from model_service.tests.fixtures.command import command  # noqa
-from model_service.tests.fixtures.data import data  # noqa
-from model_service.tests.fixtures.data_storage import data_storage  # noqa
-from model_service.tests.fixtures.dataset_info import dataset_info  # noqa
-from model_service.tests.fixtures.epoch_finished_handler_mock import epoch_finished_handler_mock  # noqa
-from model_service.tests.fixtures.feed_forward_network import feed_forward_network  # noqa
-from model_service.tests.fixtures.feed_forward_network_id import feed_forward_network_id  # noqa
-from model_service.tests.fixtures.feed_forward_network_repository import feed_forward_network_repository  # noqa
-from model_service.tests.fixtures.file_key import file_key  # noqa
-from model_service.tests.fixtures.labels_column import labels_column  # noqa
-from model_service.tests.fixtures.table_dataset_id import table_dataset_id  # noqa
-from model_service.tests.fixtures.table_dataset_repository import table_dataset_repository  # noqa
-from model_service.tests.fixtures.training_result_repository import training_result_repository  # noqa
-from model_service.tests.fixtures.training_session import training_session  # noqa
-from model_service.tests.fixtures.training_session_id import training_session_id  # noqa
-from model_service.tests.fixtures.training_session_repository import training_session_repository  # noqa
-from model_service.tests.fixtures.user import user  # noqa
