+++ tests/fixtures/training_result_repository.py
