+++ tests/fixtures/training_session_id.py
