+++ tests/fixtures/training_session_repository.py
